+++ src/inversify.config.ts
@@ -29,8 +29,6 @@ const container = new Container();
 // Intents
 const intents: GatewayIntentBits[] = [];
 intents.push(GatewayIntentBits.Guilds); // To listen for guildCreate event
-intents.push(GatewayIntentBits.GuildMessageReactions); // To listen for message reactions (messageReactionAdd event)
-intents.push(GatewayIntentBits.GuildVoiceStates); // To listen for voice state changes (voiceStateUpdate event)
 
 // Bot
 container.bind<Bot>(TYPES.Bot).to(Bot).inSingletonScope();